finish refactoring Tasks.jsx
